Services/cnn/requirements.txt
Django==2.2.4
numpy==1.15
tensorflow==1.12
matplotlib==3.1.0
scikit-image==0.14.2
google-cloud-storage
scipy